requirements.txt

streamlit==1.39.0
Pillow>=10.0.0
numpy>=1.26.0
opencv-python>=4.8.0.76
diffusers>=0.31.0
transformers>=4.44.0
accelerate>=0.33.0
safetensors>=0.4.5

# Pose (optional). If install fails on your platform, the app will still run.
mediapipe>=0.10.9
